2-MCP/llama-with-search/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
mcp
duckduckgo-search
requests